Remove useless cast

+++ examples/the-hat-chooses-the-wizard/src/level_display.rs
@@ -28,12 +28,7 @@ pub fn write_level(
     .iter()
     .enumerate()
     {
-        map.set_tile(
-            vram,
-            (i as u16, 0).into(),
-            tileset,
-            tile_settings[tile as usize],
-        );
+        map.set_tile(vram, (i as u16, 0).into(), tileset, tile_settings[tile]);
     }
 
     map.set_scroll_pos((-(WIDTH / 2 - 7 * 8 / 2) as i16, -(HEIGHT / 2 - 4) as i16).into());
